Extract work from program.cs
